# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Talk-to-Learn API
  version: 1.0.0
  description: >-
    OpenAPI-Dokumentation der Server-Endpunkte unter `app/api/*`.
    Die Spezifikation beschreibt die aktuell implementierten Request-/Response-Formate.
servers:
  - url: /
    description: Relative Basis-URL der Next.js App
security:
  - NextAuthSession: []
  - NextAuthSecureSession: []
...
paths:
  /api/progress:
    get:
      tags: [Progress]
      summary: Lernfortschritt für ein Lernset laden
      operationId: getProgress
      parameters:
        - $ref: '#/components/parameters/DeckIdQuery'
      responses:
        '200':
          description: Fortschritt geladen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      tags: [Progress]
      summary: Lernfortschritt für ein Lernset aktualisieren
      operationId: updateProgress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressPostRequest'
      responses:
        '200':
          description: Fortschritt aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    NextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
    NextAuthSecureSession:
      type: apiKey
      in: cookie
      name: __Secure-next-auth.session-token
  parameters:
    DeckIdQuery:
      in: query
      name: deckId
      required: true
      schema:
        type: string
      description: ID des Lernsets.
...
  responses:
    BadRequestError:
      description: Ungültige Anfrage
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Nicht authentifiziert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Ressource nicht gefunden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LearningStage:
      type: string
      enum: [intro, scaffolded, free]
    ReviewOutcome:
      type: string
      enum: [known, review, wrong]
    ProgressStats:
      type: object
      required: [known, learning, new]
      properties:
        known:
          type: integer
          minimum: 0
        learning:
          type: integer
          minimum: 0
        new:
          type: integer
          minimum: 0
    ProgressResponse:
      type: object
      required: [stats, nextQuestionId, learningPhase, learningStage]
      properties:
        stats:
          $ref: '#/components/schemas/ProgressStats'
        nextQuestionId:
          type: [integer, 'null']
          minimum: 0
        learningPhase:
          $ref: '#/components/schemas/LearningStage'
        learningStage:
          $ref: '#/components/schemas/LearningStage'
    ProgressPostRequest:
      oneOf:
        - $ref: '#/components/schemas/ProgressActionReset'
        - $ref: '#/components/schemas/ProgressActionMarkSeen'
        - $ref: '#/components/schemas/ProgressActionIntroReviewLater'
        - $ref: '#/components/schemas/ProgressActionScaffoldedExplained'
        - $ref: '#/components/schemas/ProgressActionSkipKnownUnknown'
        - $ref: '#/components/schemas/ProgressActionReview'
        - $ref: '#/components/schemas/ProgressActionReviewWithoutAction'
    ProgressActionBase:
      type: object
      required: [deckId, action]
      properties:
        deckId:
          type: string
        action:
          type: string
    ProgressActionReset:
      allOf:
        - $ref: '#/components/schemas/ProgressActionBase'
        - type: object
          properties:
            action:
              type: string
              enum: [reset]
    ProgressActionMarkSeen:
      allOf:
        - $ref: '#/components/schemas/ProgressActionBase'
        - type: object
          required: [questionId]
          properties:
            action:
              type: string
              enum: [mark_seen]
            questionId:
              type: integer
              minimum: 0
    ProgressActionIntroReviewLater:
      allOf:
        - $ref: '#/components/schemas/ProgressActionBase'
        - type: object
          required: [questionId]
          properties:
            action:
              type: string
              enum: [intro_review_later]
            questionId:
              type: integer
              minimum: 0
    ProgressActionScaffoldedExplained:
      allOf:
        - $ref: '#/components/schemas/ProgressActionBase'
        - type: object
          required: [questionId]
          properties:
            action:
              type: string
              enum: [scaffolded_explained]
            questionId:
              type: integer
              minimum: 0
    ProgressActionSkipKnownUnknown:
      allOf:
        - $ref: '#/components/schemas/ProgressActionBase'
        - type: object
          required: [questionId]
          properties:
            action:
              type: string
              enum: [skip_known_unknown]
            questionId:
              type: integer
              minimum: 0
    ProgressActionReview:
      allOf:
        - $ref: '#/components/schemas/ProgressActionBase'
        - type: object
          required: [questionId, outcome]
          properties:
            action:
              type: string
              enum: [review]
            questionId:
              type: integer
              minimum: 0
            outcome:
              $ref: '#/components/schemas/ReviewOutcome'
    ProgressActionReviewWithoutAction:
      type: object
      required: [deckId, questionId, outcome]
      properties:
        deckId:
          type: string
        questionId:
          type: integer
          minimum: 0
        outcome:
          $ref: '#/components/schemas/ReviewOutcome'
...